requirements.txt
Click==7.0
github3.py
tqdm
gitpython
